FilteringZipDependencySelector and test
